backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6
python-jose[cryptography]  # For JWT
passlib[bcrypt]
google-generativeai
python-dotenv
